refactor node replacement logic
Moving the logic used to replace node types for a builder into its own
source file. This is to help minimize the existing and future size of
the main builder's implementation.
